requirements.txt
setuptools
numpy
PILLOW
picamera[array]